Move 'getDiagnostic' method to BCompileUtil class
