+++ modules/launcher/src/main/java/org/ballerinalang/launcher/util/BCompileUtil.java
@@ -24,6 +24,7 @@ import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.util.codegen.PackageInfo;
 import org.ballerinalang.util.codegen.ProgramFile;
 import org.ballerinalang.util.codegen.StructInfo;
+import org.ballerinalang.util.diagnostic.Diagnostic;
 import org.ballerinalang.util.diagnostic.DiagnosticListener;
 import org.wso2.ballerinalang.compiler.Compiler;
 import org.wso2.ballerinalang.compiler.tree.BLangPackage;
@@ -305,5 +306,41 @@ public class BCompileUtil {
         StructInfo structInfo = structPackageInfo.getStructInfo(structName);
         BStructType structType = structInfo.getType();
         return new BStruct(structType);
+    }
+
+
+    /**
+     * Used by IntelliJ IDEA plugin to provide semantic analyzing capability.
+     *
+     * @param classLoader a {@link ClassLoader} to be set as thread context class loader. This is used by {@link
+     *                    java.util.ServiceLoader}. Otherwise semantic analyzing capability providing wont work since it
+     *                    cant find core package.
+     * @param sourceRoot  source root of a project
+     * @param fileName    either the file name (if in project root) or the package name
+     * @return list of diagnostics
+     */
+    public static List<Diagnostic> getDiagnostics(ClassLoader classLoader, String sourceRoot, String fileName) {
+        Thread.currentThread().setContextClassLoader(classLoader);
+        CompilerContext context = new CompilerContext();
+        CompilerOptions options = CompilerOptions.getInstance(context);
+        options.put(SOURCE_ROOT, sourceRoot);
+        options.put(COMPILER_PHASE, CompilerPhase.CODE_GEN.toString());
+        options.put(PRESERVE_WHITESPACE, "false");
+
+        CompileResult comResult = new CompileResult();
+
+        // catch errors
+        DiagnosticListener listener = comResult::addDiagnostic;
+        context.put(DiagnosticListener.class, listener);
+
+        // compile
+        Compiler compiler = Compiler.getInstance(context);
+        compiler.compile(fileName);
+        org.wso2.ballerinalang.programfile.ProgramFile programFile = compiler.getCompiledProgram();
+        if (programFile != null) {
+            comResult.setProgFile(LauncherUtils.getExecutableProgram(programFile));
+        }
+        Diagnostic[] diagnostics = comResult.getDiagnostics();
+        return Arrays.stream(diagnostics).collect(Collectors.toList());
     }
 }
